Tested pre-processor.c. Made everything work, so changes in the entire project's scope were necessary. Many small changes in FileHandling, NameTable, Errors folder (made the project ignore the errors from not importing everything off of mmn-22), but mostly in general_help_methods.c.
Everything is working correctly after the changes, but need to add more documentations on the added parts and work with valgrind to fix memory issues.

(Sunday, 25/06/23)

+++ FileHandling/readFromFile.h
@@ -18,7 +18,7 @@
  * @param   **line is a pointer to string that will hold the read line.
  * @return  The length of the line if it has been read successfully or EOF (-1) if not.
  */
-int readNextLineFromFile(const char *file_name, const char *fileType, char **line);
+ssize_t readNextLineFromFile(const char *file_name, const char *fileType, char **line);
 
 /* ---------------------------------------- */
